requirements.txt
# Dépendances supplémentaires
Cython

# Dépendances principales
prophet
pandas==1.5.3
numpy
joblib

# Dépendances pour l'analyse et la visualisation
matplotlib
plotly==5.18.0

# Dépendances pour l'API (si nécessaire)
fastapi
pydantic
uvicorn
starlette
aiofiles
xgboost

# Dépendances pour l'historique et l'export
sqlalchemy==1.4.41
openpyxl
reportlab==3.6.12
scikit-learn

# Dépendances utilitaires
python-dotenv
pystan
python-multipart

# Dépendances pour la gestion de bases de données
mysql-connector-python
pymysql
redis
fastapi-cache2